integration tests: use local test-util for dockerd
While the upstream docker testutil could be useful, currently buildkit
only uses a minimum subset of that package (i.e., it only starts a
daemon, stops it, and collects logs).

This patch creates a local, bare minimum implementation of the testutil
for use in the integration tests to reduce the dependency graph.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>