cache: Refactor out cacheRecord equal* fields
The equalMutable and equalImmutable fields were used to implement the
"lazy commit" feature (not to be confused with lazy refs), which allows
refs to put off committing their snapshots to the underlying driver until
absolutely needed.

However, this way of modeling it resulted in there being 2 actual
cacheRecords even when the underlying snapshot was the same. This also
lead to a lot of convoluted logic checking for the existence of the
equal* fields.

This commits switches to a new approach of only maintaining a single
cacheRecord for what used to be separate equal* records. The cacheRecord
internally keeps track of whether it has been finalized or not and
updates which snapshot ID is underlying it.

This results in overall much simpler logic, as can be seen by looking at
the before+after of GetMutable, Commit, prune, and release, among
others. The only meaningfully-sized block of new code added is in
cacheManager.init, which handles migrating records from the old format
and can thus eventually be deleted once the older versions of buildkit
are no longer supported.

This change is backwards compatible because of that migration logic in
cacheManager.init and because MutableRefs have never been saved in the
upper-level solver cache, only ImmutableRefs are there, so removing any
refs that were marked as an equalMutable and merging them into their
equalImmutable will not result in any cache keys unexpectedly missing.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>