Fixes Windows symlinks targets
When building with the output type=oci or type=registry, for Windows
symlinks, "Files/" is prepended to the symlink Target, which can break
the symlink entirely if it's an absolute path (C:\path\to\file becomes
Files\C:\path\to\file), and for relative paths workarounds were needed
(path\to\file would become Files\path\to\file, and it would have worked
if that path was valid).

This problem does not occur when the output type is tar.

Removes the "Files/" prefix from the Windows symlinks.